refactored app name and app cart integration
