+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/afriprize2.iml" filepath="$PROJECT_DIR$/.idea/afriprize2.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/easyph.iml" filepath="$PROJECT_DIR$/.idea/easyph.iml" />
     </modules>
   </component>
 </project>
