+++ android/app/src/main/kotlin/com/echobitstech/easyph/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.afriprize.afriprize
+package com.echobitstech.easyph
 
 import io.flutter.embedding.android.FlutterActivity
